# Conflicts:
#	code/__HELPERS/_macros.dm
#	code/game/mecha/equipment/weapons/weapons.dm

+++ code/game/gamemodes/gameticker.dm
@@ -1,18 +1,18 @@
 var/datum/controller/gameticker/ticker
 
-#define GAME_STATE_PREGAME		1
-#define GAME_STATE_SETTING_UP	2
-#define GAME_STATE_PLAYING		3
-#define GAME_STATE_FINISHED		4
-
-
 /datum/controller/gameticker
 	var/remaining_time = 0
-	var/const/restart_timeout = 600
+	var/const/restart_timeout = 60 SECONDS //Right now, this is padded out by the end credit's audio starting time (at the time of writing this, 10 seconds)
 	var/current_state = GAME_STATE_PREGAME
+	var/gamestart_time = -1 //In seconds. Set by ourselves in setup()
+	var/shuttledocked_time = -1 //In seconds. Set by emergency_shuttle/proc/shuttle_phase()
+	var/gameend_time = -1 //In seconds. Set by ourselves in process()
+
+	var/pregame_timeleft = 0
+	var/delay_end = 0	//if set to nonzero, the round will not restart on its own
 
 	var/hide_mode = 0
-	var/datum/game_mode/mode = null
+	var/datum/gamemode/mode = null
 	var/event_time = null
 	var/event = 0
 
@@ -20,25 +20,25 @@ var/datum/controller/gameticker/ticker
 
 	var/list/datum/mind/minds = list()//The people in the game. Used for objective tracking.
 
-	var/Bible_icon_state	// icon_state the chaplain has chosen for his bible
-	var/Bible_item_state	// item_state the chaplain has chosen for his bible
+	var/Bible_icon_state	// icon_state the OFFICIAL chaplain has chosen for his bible
+	var/Bible_item_state	// item_state the OFFICIAL chaplain has chosen for his bible
 	var/Bible_name			// name of the bible
-	var/Bible_deity_name = "Space Jesus"
+	var/Bible_deity_name = "Space Jesus" 	// Default deity
+	var/datum/religion/chap_rel 			// Official religion of chappy
+	var/list/datum/religion/religions = list() // Religion(s) in the game
 
 	var/random_players = 0 	// if set to nonzero, ALL players who latejoin or declare-ready join will have random appearances/genders
 
 	var/hardcore_mode = 0	//If set to nonzero, hardcore mode is enabled (current hardcore mode features: damage from hunger)
 							//Use the hardcore_mode_on macro - if(hardcore_mode_on) to_chat(user,"You're hardcore!")
+	var/datum/rune_controller/rune_controller
 
-	var/list/syndicate_coalition = list() // list of traitor-compatible factions
-	var/list/factions = list()			  // list of all factions
-	var/list/availablefactions = list()	  // list of factions with openings
+	var/triai = 0 //Global holder for Triumvirate
 
-	var/pregame_timeleft = 0
+	var/explosion_in_progress
+	var/station_was_nuked
 
-	var/delay_end = 0	//if set to nonzero, the round will not restart on it's own
-
-	var/triai = 0//Global holder for Triumvirate
+	var/list/datum/role/antag_types = list() // Associative list of all the antag types in the round (List[id] = roleNumber1) //Seems to be totally unused?
 
 	// Hack
 	var/obj/machinery/media/jukebox/superjuke/thematic/theme = null
@@ -63,11 +63,27 @@ var/datum/controller/gameticker/ticker
 		"sound/music/starman.ogg",
 		"sound/music/dawsonschristian.ogg",
 		"sound/music/carmenmirandasghost.ogg",
+		"sound/music/twilight.ogg",
 		))
-	login_music = fcopy_rsc(oursong)
+
+	if(SNOW_THEME)
+		var/path = "sound/music/xmas/"
+		var/list/filenames = flist(path)
+		for(var/filename in filenames)
+			if(copytext(filename, length(filename)) == "/")
+				filenames -= filename
+		login_music = file("[path][pick(filenames)]")
+	else
+		login_music = fcopy_rsc(oursong)
+
+	send2maindiscord("**Server is loaded** and in pre-game lobby at `[config.server? "byond://[config.server]" : "byond://[world.address]:[world.port]"]`")
 
 	do
-		var/delay_timetotal = 3000 //actually 5 minutes or incase this is changed from 3000, (time_in_seconds * 10)
+#ifdef GAMETICKER_LOBBY_DURATION
+		var/delay_timetotal = GAMETICKER_LOBBY_DURATION
+#else
+		var/delay_timetotal = DEFAULT_LOBBY_TIME
+#endif
 		pregame_timeleft = world.timeofday + delay_timetotal
 		to_chat(world, "<B><FONT color='blue'>Welcome to the pre-game lobby!</FONT></B>")
 		to_chat(world, "Please, setup your character and select ready. Game will start in [(pregame_timeleft - world.timeofday) / 10] seconds.")
@@ -112,15 +128,15 @@ var/datum/controller/gameticker/ticker
 	//Create and announce mode
 	if(master_mode=="secret")
 		src.hide_mode = 1
-	var/list/datum/game_mode/runnable_modes
-	if((master_mode=="random") || (master_mode=="secret"))
+	var/list/datum/gamemode/runnable_modes
+	if((master_mode=="random"))
 		runnable_modes = config.get_runnable_modes()
 		if (runnable_modes.len==0)
 			current_state = GAME_STATE_PREGAME
 			to_chat(world, "<B>Unable to choose playable game mode.</B> Reverting to pre-game lobby.")
 			return 0
 		if(secret_force_mode != "secret")
-			var/datum/game_mode/M = config.pick_mode(secret_force_mode)
+			var/datum/gamemode/M = config.pick_mode(secret_force_mode)
 			if(M.can_start())
 				src.mode = config.pick_mode(secret_force_mode)
 		job_master.ResetOccupations()
@@ -129,10 +145,15 @@ var/datum/controller/gameticker/ticker
 		if(src.mode)
 			var/mtype = src.mode.type
 			src.mode = new mtype
+	else if (master_mode=="secret")
+		mode = config.pick_mode("Dynamic Mode") //Huzzah
 	else
 		src.mode = config.pick_mode(master_mode)
+
+	//log_startup_progress("gameticker.mode is [src.mode.name].")
+	src.mode = new mode.type
 	if (!src.mode.can_start())
-		to_chat(world, "<B>Unable to start [mode.name].</B> Not enough players, [mode.required_players] players needed. Reverting to pre-game lobby.")
+		to_chat(world, "<B>Unable to start [mode.name].</B> Not enough players, [mode.minimum_player_count] players needed. Reverting to pre-game lobby.")
 		del(mode)
 		current_state = GAME_STATE_PREGAME
 		job_master.ResetOccupations()
@@ -140,7 +161,7 @@ var/datum/controller/gameticker/ticker
 
 	//Configure mode and assign player to special mode stuff
 	job_master.DivideOccupations() //Distribute jobs
-	var/can_continue = src.mode.pre_setup()//Setup special modes
+	var/can_continue = src.mode.Setup()//Setup special modes
 	if(!can_continue)
 		current_state = GAME_STATE_PREGAME
 		to_chat(world, "<B>Error setting up [master_mode].</B> Reverting to pre-game lobby.")
@@ -152,30 +173,41 @@ var/datum/controller/gameticker/ticker
 
 	if(hide_mode)
 		var/list/modes = new
-		for (var/datum/game_mode/M in runnable_modes)
+		for (var/datum/gamemode/M in runnable_modes)
 			modes+=M.name
 		modes = sortList(modes)
-		to_chat(world, "<B>The current game mode is - Secret!</B>")
-		to_chat(world, "<B>Possibilities:</B> [english_list(modes)]")
-	else
-		src.mode.announce()
+		if(Holiday == APRIL_FOOLS_DAY)
+			to_chat(world, "<B>The current game mode is - [pick("Chivalry","Crab Battle","Bay Transfer","Dwarf Fortress","Ian Says","Admins Funhouse","Meteor","Xenoarchaeology Appreciation","Clowns versus [pick("Mimes","Assistants","the Universe")]","Dino wars","Malcolm in the Middle","Six hours of extended where one person with all the access refuses to call the shuttle while everyone else goes braindead","Monkey Study","Nations","Nations by Hasbro","High roleplay Extended","DarkRP","Babies Day out","Ians Day out","Shortstaffed medical")]!</B>")
+		else
+			to_chat(world, "<B>The current game mode is - Secret!</B>")
+			to_chat(world, "<B>Possibilities:</B> [english_list(modes)]")
+
+	gamestart_time = world.time / 10
 
 	init_PDAgames_leaderboard()
 	create_characters() //Create player characters and transfer them
 	collect_minds()
 	equip_characters()
-	data_core.manifest()
 	current_state = GAME_STATE_PLAYING
+
+	// Update new player panels so they say join instead of ready up.
+	for(var/mob/new_player/player in player_list)
+		player.new_player_panel_proc()
+
 
 	//here to initialize the random events nicely at round start
 	setup_economy()
 
+#if UNIT_TESTS_AUTORUN
+	run_unit_tests()
+#endif
+
 	spawn(0)//Forking here so we dont have to wait for this to finish
-		mode.post_setup()
+		mode.PostSetup()
 		//Cleanup some stuff
 		for(var/obj/effect/landmark/start/S in landmarks_list)
-			//Deleting Startpoints but we need the ai point to AI-ize people later
-			if (S.name != "AI")
+			//Deleting Startpoints but we need the ai point to AI-ize people later and the Trader point to throw new ones
+			if (S.name != "AI" && S.name != "Trader")
 				qdel(S)
 		var/list/obj/effect/landmark/spacepod/random/L = list()
 		for(var/obj/effect/landmark/spacepod/random/SS in landmarks_list)
@@ -188,8 +220,10 @@ var/datum/controller/gameticker/ticker
 				if(istype(obj, /obj/effect/landmark/spacepod/random))
 					qdel(obj)
 
-		stat_collection.death_stats = list() // Get rid of the corpses that spawn on startup.
 		to_chat(world, "<FONT color='blue'><B>Enjoy the game!</B></FONT>")
+
+		send2maindiscord("**The game has started**")
+
 //		world << sound('sound/AI/welcome.ogg')// Skie //Out with the old, in with the new. - N3X15
 
 		if(!config.shut_up_automatic_diagnostic_and_announcement_system)
@@ -203,12 +237,18 @@ var/datum/controller/gameticker/ticker
 				'sound/AI/vox_reminder6.ogg',
 				'sound/AI/vox_reminder7.ogg',
 				'sound/AI/vox_reminder8.ogg',
-				'sound/AI/vox_reminder9.ogg')
+				'sound/AI/vox_reminder9.ogg',
+				'sound/AI/vox_reminder10.ogg',
+				'sound/AI/vox_reminder11.ogg',
+				'sound/AI/vox_reminder12.ogg',
+				'sound/AI/vox_reminder13.ogg',
+				'sound/AI/vox_reminder14.ogg',
+				'sound/AI/vox_reminder15.ogg')
 			for(var/sound in welcome_sentence)
 				play_vox_sound(sound,STATION_Z,null)
 		//Holiday Round-start stuff	~Carn
 		Holiday_Game_Start()
-		mode.Clean_Antags()
+		//mode.Clean_Antags()
 
 	//start_events() //handles random events and space dust.
 	//new random event system is handled from the MC.
@@ -228,7 +268,7 @@ var/datum/controller/gameticker/ticker
 		population_poll_loop()
 
 	wageSetup()
-
+	post_roundstart()
 	return 1
 
 /datum/controller/gameticker
@@ -336,10 +376,20 @@ var/datum/controller/gameticker/ticker
 			if(player.mind.assigned_role=="AI")
 				player.close_spawn_windows()
 				player.AIize()
+			else if(player.mind.assigned_role=="Cyborg")
+				player.create_roundstart_cyborg()
+
 			else if(!player.mind.assigned_role)
 				continue
 			else
-				player.FuckUpGenes(player.create_character())
+
+				var/mob/living/carbon/human/new_character = player.create_character()
+				switch(new_character.mind.assigned_role)
+					if("MODE","Mobile MMI","Trader")
+						//No injection
+					else
+						data_core.manifest_inject(new_character)
+				player.FuckUpGenes(new_character)
 				qdel(player)
 
 
@@ -386,14 +436,18 @@ var/datum/controller/gameticker/ticker
 		force_round_end=1
 
 	var/mode_finished = mode.check_finished() || (emergency_shuttle.location == 2 && emergency_shuttle.alert == 1) || force_round_end
-	if(!mode.explosion_in_progress && mode_finished)
+	if(!explosion_in_progress && mode_finished)
 		current_state = GAME_STATE_FINISHED
 
 		spawn
 			declare_completion()
+
+			end_credits.on_round_end()
+
+			gameend_time = world.time / 10
 			if(config.map_voting)
 				//testing("Vote picked [chosen_map]")
-				vote.initiate_vote("map","The Server", popup = 1, weighted_vote = 1)
+				vote.initiate_vote("map","The Server", popup = 1, weighted_vote = config.weighted_votes)
 				var/options = jointext(vote.choices, " ")
 				feedback_set("map vote choices", options)
 
@@ -408,9 +462,8 @@ var/datum/controller/gameticker/ticker
 				log_game("Server chose [watchdog.chosen_map]!")
 
 
-
 		spawn(50)
-			if (mode.station_was_nuked)
+			if (station_was_nuked)
 				feedback_set_details("end_proper","nuke")
 				if(!delay_end && !watchdog.waiting)
 					to_chat(world, "<span class='notice'><B>Rebooting due to destruction of station in [restart_timeout/10] seconds</B></span>")
@@ -425,6 +478,8 @@ var/datum/controller/gameticker/ticker
 						blackbox.save_all_data_to_sql()
 				else
 					blackbox.save_all_data_to_sql()
+
+			stat_collection.Process()
 
 			if (watchdog.waiting)
 				to_chat(world, "<span class='notice'><B>Server will shut down for an automatic update in [config.map_voting ? "[(restart_timeout/10)] seconds." : "a few seconds."]</B></span>")
@@ -448,12 +503,6 @@ var/datum/controller/gameticker/ticker
 				delay_end = 2
 
 	return 1
-
-/datum/controller/gameticker/proc/getfactionbyname(var/name)
-	for(var/datum/faction/F in factions)
-		if(F.name == name)
-			return F
-
 
 /datum/controller/gameticker/proc/init_PDAgames_leaderboard()
 	init_snake_leaderboard()
@@ -482,62 +531,17 @@ var/datum/controller/gameticker/ticker
 	minesweeper_best_players["expert"] = "none"
 
 /datum/controller/gameticker/proc/declare_completion()
-	var/ai_completions = "<h1>Round End Information</h1><HR>"
-
-	ai_completions += "<h3>Silicons Laws</h3>"
-	for(var/mob/living/silicon/ai/ai in mob_list)
-		var/icon/flat = getFlatIcon(ai)
-		end_icons += flat
-		var/tempstate = end_icons.len
-		if(ai.stat != 2)
-			ai_completions += {"<br><b><img src="logo_[tempstate].png"> [ai.name] (Played by: [ai.key])'s laws at the end of the game were:</b>"}
-		else
-			ai_completions += {"<br><b><img src="logo_[tempstate].png"> [ai.name] (Played by: [ai.key])'s laws when it was deactivated were:</b>"}
-		ai_completions += "<br>[ai.write_laws()]"
-
-		if (ai.connected_robots.len)
-			var/robolist = "<br><b>The AI's loyal minions were:</b> "
-			for(var/mob/living/silicon/robot/robo in ai.connected_robots)
-				if (!robo.connected_ai || !isMoMMI(robo)) // Don't report MoMMIs or unslaved robutts
-					continue
-				robolist += "[robo.name][robo.stat?" (Deactivated) (Played by: [robo.key]), ":" (Played by: [robo.key]), "]"
-			ai_completions += "[robolist]"
-
-	for (var/mob/living/silicon/robot/robo in mob_list)
-		if(!robo)
-			continue
-		var/icon/flat = getFlatIcon(robo)
-		end_icons += flat
-		var/tempstate = end_icons.len
-		if (!robo.connected_ai)
-			if (robo.stat != 2)
-				ai_completions += {"<br><b><img src="logo_[tempstate].png"> [robo.name] (Played by: [robo.key]) survived as an AI-less [isMoMMI(robo)?"MoMMI":"borg"]! Its laws were:</b>"}
-			else
-				ai_completions += {"<br><b><img src="logo_[tempstate].png"> [robo.name] (Played by: [robo.key]) was unable to survive the rigors of being a [isMoMMI(robo)?"MoMMI":"cyborg"] without an AI. Its laws were:</b>"}
-		else
-			ai_completions += {"<br><b><img src="logo_[tempstate].png"> [robo.name] (Played by: [robo.key]) [robo.stat!=2?"survived":"perished"] as a [isMoMMI(robo)?"MoMMI":"cyborg"] slaved to [robo.connected_ai]! Its laws were:</b>"}
-		ai_completions += "<br>[robo.write_laws()]"
-
-	for(var/mob/living/silicon/pai/pAI in mob_list)
-		var/icon/flat
-		flat = getFlatIcon(pAI)
-		end_icons += flat
-		var/tempstate = end_icons.len
-		ai_completions += {"<br><b><img src="logo_[tempstate].png"> [pAI.name] (Played by: [pAI.key]) [pAI.stat!=2?"survived":"perished"] as a pAI whose master was [pAI.master]! Its directives were:</b><br>[pAI.write_directives()]"}
-
-	mode.declare_completion()//To declare normal completion.
-
-	ai_completions += "<HR><BR><h2>Mode Result</h2>"
-	//ai_completions += "<br>[mode.completion_text]"
-
-	scoreboard(ai_completions)
-
+	if(!ooc_allowed)
+		to_chat(world, "<B>The OOC channel has been automatically re-enabled!</B>")
+		ooc_allowed = TRUE
+	scoreboard()
 	return 1
 
+/*
 /datum/controller/gameticker/proc/ert_declare_completion()
 	var/text = ""
 	if( ticker.mode.ert.len )
-		var/icon/logo = icon('icons/mob/mob.dmi', "ert-logo")
+		var/icon/logo = icon('icons/logos.dmi', "ert-logo")
 		end_icons += logo
 		var/tempstate = end_icons.len
 		text += {"<br><img src="logo_[tempstate].png"> <FONT size = 2><B>The emergency responders were:</B></FONT> <img src="logo_[tempstate].png">"}
@@ -569,7 +573,7 @@ var/datum/controller/gameticker/ticker
 /datum/controller/gameticker/proc/deathsquad_declare_completion()
 	var/text = ""
 	if( ticker.mode.deathsquad.len )
-		var/icon/logo = icon('icons/mob/mob.dmi', "death-logo")
+		var/icon/logo = icon('icons/logos.dmi', "death-logo")
 		end_icons += logo
 		var/tempstate = end_icons.len
 		text += {"<br><img src="logo_[tempstate].png"> <FONT size = 2><B>The death commando were:</B></FONT> <img src="logo_[tempstate].png">"}
@@ -597,7 +601,7 @@ var/datum/controller/gameticker/ticker
 		text += "<BR><HR>"
 
 	return text
-
+*/
 /datum/controller/gameticker/proc/bomberman_declare_completion()
 	var/icon/bomberhead = icon('icons/obj/clothing/hats.dmi', "bomberman")
 	end_icons += bomberhead
@@ -694,6 +698,30 @@ var/datum/controller/gameticker/ticker
 		text += {"<br><img src="logo_[tempstate].png"> [winner]"}
 
 	return text
+
+/datum/controller/gameticker/proc/get_all_heads()
+	var/list/heads = list()
+	for(var/mob/player in mob_list)
+		if(player.mind && (player.mind.assigned_role in command_positions))
+			heads += player.mind
+	return heads
+
+/datum/controller/gameticker/proc/get_assigned_head_roles()
+	var/list/roles = list()
+	for(var/mob/player in mob_list)
+		if(player.mind && (player.mind.assigned_role in command_positions))
+			roles += player.mind.assigned_role
+	return roles
+
+/datum/controller/gameticker/proc/post_roundstart()
+	//Handle all the cyborg syncing
+	var/list/active_ais = active_ais()
+	if(active_ais.len)
+		for(var/mob/living/silicon/robot/R in cyborg_list)
+			if(!R.connected_ai)
+				R.connect_AI(select_active_ai_with_fewest_borgs())
+				to_chat(R, R.connected_ai?"<b>You have synchronized with an AI. Their name will be stated shortly. Other AIs can be ignored.</b>":"<b>You are not synchronized with an AI, and therefore are not required to heed the instructions of any unless you are synced to them.</b>")
+			R.lawsync()
 
 
 /world/proc/has_round_started()
